+++ packages/zendesk-adapter/src/filters/user.ts
@@ -17,7 +17,7 @@ import _ from 'lodash'
 import { Change, InstanceElement, getChangeData, isInstanceElement } from '@salto-io/adapter-api'
 import { client as clientUtils } from '@salto-io/adapter-components'
 import { FilterCreator } from '../filter'
-import { TYPE_NAME_TO_REPLACER, getIdByEmail } from '../user_utils'
+import { TYPE_NAME_TO_REPLACER, getIdByEmail } from '../users/user_utils'
 import { deployModificationFunc } from '../replacers_utils'
 import { paginate } from '../client/pagination'
 import { FETCH_CONFIG } from '../config'
