+++ packages/zendesk-adapter/test/filters/audit_logs.test.ts
@@ -28,7 +28,7 @@ import ZendeskClient from '../../src/client/client'
 import filterCreator, { AUDIT_TIME_TYPE_ID, DELETED_USER } from '../../src/filters/audit_logs'
 import { createFilterCreatorParams } from '../utils'
 import { DEFAULT_CONFIG, FETCH_CONFIG } from '../../src/config'
-import { getIdByName } from '../../src/user_utils'
+import { getIdByName } from '../../src/users/user_utils'
 import {
   ARTICLE_TRANSLATION_TYPE_NAME,
   AUDIT_TIME_TYPE_NAME,
